requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
Django==3.2.5
greenlet==1.1.0
image==1.5.33
Pillow==8.3.1
psycopg2-binary==2.9.1
python-telegram-bot==13.7
pytz==2021.1
qrcode==7.1
six==1.16.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
tornado==6.1
transliterate==1.10.2
tzlocal==2.1
